requirements.txt
flask
google-generativeai
chromadb
